requirements.txt
openai 
azure-identity 
azure-ai-projects==2.0.0b1
jsonref
python-dotenv